+++ src/modules/api.ts
@@ -5,6 +5,7 @@ import { Syncs, SyncsAPI } from './syncs';
 import { Integrations } from './integrations';
 import { Webhooks } from './webhooks';
 import { DataStores } from './datastores';
+import { Issues } from './issues';
 
 export class API {
     auth: AuthType;
@@ -15,6 +16,7 @@ export class API {
     Integrations;
     Webhooks;
     DataStores;
+    Issues;
 
     constructor(auth: AuthType) {
         this.auth = auth;
@@ -24,6 +26,7 @@ export class API {
         this.Integrations = Integrations(this.internalApi);
         this.Webhooks = Webhooks(this.internalApi);
         this.DataStores = DataStores(this.internalApi);
+        this.Issues = Issues(this.internalApi);
     }
 
     private _setup = async () => {