+++ src/modules/ecommerce.ts
@@ -46,7 +46,9 @@ const ecommerceFactory = {
     getProducts(
         params?: GetProductsParams,
         options?: RawDataOption
-    ): RequestData<components['schemas']['ProductItem-Output'][]> {
+    ): RequestData<
+        components['schemas']['backbone_common__models__commerce__common__ProductItem'][]
+    > {
         return {
             params,
             method: 'get',
@@ -67,7 +69,9 @@ const ecommerceFactory = {
     getProduct(
         productId: string,
         options?: RawDataOption
-    ): RequestData<components['schemas']['ProductItem-Output']> {
+    ): RequestData<
+        components['schemas']['backbone_common__models__commerce__common__ProductItem']
+    > {
         return {
             method: 'get',
             url: `/consumers/{consumer_id}/commerce/products/${productId}`,